pyproject.toml
[project]
name = "lenggraph-experemenations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.11.0",
    "firecrawl-py>=4.8.0",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph-cli[inmem]>=0.4.7",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "pydantic>=2.12.4",
    "qdrant-client>=1.15.1",
]
